improve IAM public policies logic
