+++ prowler/providers/aws/services/sqs/sqs_queues_not_publicly_accessible/sqs_queues_not_publicly_accessible.py
@@ -1,7 +1,5 @@
 from prowler.lib.check.models import Check, Check_Report_AWS
-from prowler.providers.aws.lib.policy_condition_parser.policy_condition_parser import (
-    is_condition_block_restrictive,
-)
+from prowler.providers.aws.services.iam.lib.policy import is_condition_block_restrictive
 from prowler.providers.aws.services.sqs.sqs_client import sqs_client
